weather 파일 정상화
