+++ client/src/components/weather/css/WeatherMain.module.css
@@ -1,0 +1,129 @@
+.container {
+    grid-column: 1;
+    grid-row: 1;
+    display: grid;
+    grid-template-columns: 11fr 36fr 3fr;
+    grid-template-rows: 11fr 50fr;
+}
+
+.card {
+    grid-column: 2;
+    grid-row: 2;
+    width: 100%;
+    height: 100%;
+    display: grid;
+    grid-template-columns: 1fr;
+    grid-template-rows: 15fr 21fr 14fr;
+    border-radius: 50px;
+    font-Family: Roboto sans-serif;
+    font-Weight: 500;
+}
+
+
+/* 아이콘 스타일 */
+.icon {
+    grid-column: 1;
+    grid-row: 1;
+    width: 100%;
+    height: 100%;
+    display: flex;
+    justify-content: center;
+    align-items: flex-end;
+    user-select: none;
+}
+
+.icon img {
+    width: 136px;
+    height: 136px;
+}
+
+@media (max-width: 1920px) and (max-height: 1080px) {
+    .icon img {
+        width: 110px;
+        height: 110px;
+    }
+}
+
+
+
+/*temp 스타일*/
+.temp {
+    grid-column: 1;
+    grid-row: 2;
+    display: grid;
+    grid-template-columns: 1fr;
+    grid-template-rows: 17fr 4fr;
+}
+
+.temp div:nth-child(1) {
+    grid-column: 1;
+    grid-row: 1;
+    display: flex;
+    justify-content: center;
+    margin: 0;
+    font-size: 120px;
+    text-align: center;
+}
+
+.temp div:nth-child(1) sup {
+    font-size: 0.25em;
+    vertical-align: super;
+}
+
+.temp div:nth-child(2) {
+    font-size: 30px;
+    font-weight: 500;
+    grid-column: 1;
+    grid-row: 2;
+    display: flex;
+    justify-content: center;
+    font-Family: Roboto sans-serif;
+    font-Weight: 500;
+}
+
+@media (max-width: 1920px) and (max-height: 1080px) {
+    .temp div:nth-child(1) {
+        font-size: 100px;
+    }
+
+    .temp div:nth-child(2) {
+        font-size: 25px;
+    }
+}
+
+
+/* see more 버튼 스타일 */
+.button {
+    grid-column: 1;
+    grid-row: 3;
+    align-self: end;
+    justify-self: center;
+    width: 96%;
+    height: 30%;
+    border-radius: 0 0 50px 50px;
+    border: none;
+    display: flex;
+    justify-content: space-between;
+    background-color: transparent;
+    padding: 0 30px;
+    color: #7a7a7a;
+    font-Family: Roboto sans-serif;
+    font-Weight: 500;
+}
+
+@media (max-width: 1920px) and (max-height: 1080px) {
+    .button {
+        font-size: 12px;
+    }
+}
+
+.button:hover {
+    transition: color 0.3s ease-in-out;
+    color: #000;
+    /* Hover 시 텍스트 색 변경 */
+}
+
+.button:not(:hover) {
+    transition: color 0.3s ease-in-out;
+    /* 원래 색으로 돌아갈 때 속도 변경 */
+}
